Added RAJA atomic wrappers for tensor and vector geometry classes, fixed type discrepancy in artificial viscosity classes, added basic exec policy to config that is subject to improvement, started porting the eval deriv in SPH.cc

+++ src/ArtificialViscosity/LimitedMonaghanGingoldViscosityViewInline.hh
@@ -90,8 +90,8 @@ void
 LimitedMonaghanGingoldViscosityView<Dimension>::
 QPiij(Scalar& QPiij, Scalar& QPiji,      // result for QPi (Q/rho^2)
       Scalar& Qij, Scalar& Qji,          // result for viscous pressure
-      const unsigned nodeListi, const unsigned i, 
-      const unsigned nodeListj, const unsigned j,
+      const size_t nodeListi, const size_t i,
+      const size_t nodeListj, const size_t j,
       const Vector& xi,
       const SymTensor& Hi,
       const Vector& etai,
@@ -107,7 +107,6 @@ QPiij(Scalar& QPiij, Scalar& QPiji,      // result for QPi (Q/rho^2)
       const FieldListView<Dimension, Scalar>& fCl,
       const FieldListView<Dimension, Scalar>& fCq,
       const FieldListView<Dimension, Tensor>& DvDx) const {
-
   // Preconditions
   REQUIRE(fCl.size() == fCq.size());
   REQUIRE(DvDx.size() > std::max(nodeListi, nodeListj));
